refactor(base): move all send-queue related types to a new store::send_queue module
No changes in functionality, only code motion.
